creating new file for tests
